refactor(agent-task): split runner policy module
