+++ src/core/agent_task_loop_controller.rs
@@ -7,6 +7,14 @@ use std::io::ErrorKind;
 use std::path::PathBuf;
 
 use crate::core::agent_task_lifecycle::AgentTaskRunState;
+use crate::core::agent_task_loop_runner_policy::{
+    blocked_runner_decision, runner_policy_for_action,
+};
+pub use crate::core::agent_task_loop_runner_policy::{
+    AgentTaskLoopLocalFallbackPolicy, AgentTaskLoopRunnerAvailability,
+    AgentTaskLoopRunnerExecutionTarget, AgentTaskLoopRunnerPolicy,
+    AgentTaskLoopRunnerPolicyDecision,
+};
 use crate::core::{agent_task_lifecycle, paths, Error, Result};
 
 pub const AGENT_TASK_LOOP_CONTROLLER_SCHEMA: &str = "homeboy/agent-task-loop-controller/v1";
@@ -462,41 +470,6 @@ pub enum AgentTaskLoopActionStatus {
     BlockedRunnerUnavailable,
     BlockedRemoteMaterialization,
     BlockedLocalFallbackDenied,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-pub struct AgentTaskLoopRunnerPolicy {
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub runner: Option<String>,
-    #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub local_fallback: Option<AgentTaskLoopLocalFallbackPolicy>,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum AgentTaskLoopLocalFallbackPolicy {
-    Allowed,
-    Denied,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum AgentTaskLoopRunnerAvailability {
-    Available,
-    Unavailable { reason: String },
-    MaterializationBlocked { reason: String },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum AgentTaskLoopRunnerExecutionTarget {
-    Local,
-    Runner(String),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct AgentTaskLoopRunnerPolicyDecision {
-    pub target: Option<AgentTaskLoopRunnerExecutionTarget>,
-    pub blocked_status: Option<AgentTaskLoopActionStatus>,
-    pub diagnostic: Option<AgentTaskLoopActionDiagnostic>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -1042,27 +1015,6 @@ impl AgentTaskLoopControllerRecord {
         Ok(())
     }
 
-    fn runner_policy_for_action(
-        &self,
-        action: &AgentTaskLoopPolicyAction,
-    ) -> AgentTaskLoopRunnerPolicy {
-        let request = action_runner_request(action);
-        let runner = request
-            .and_then(|value| value.get("runner"))
-            .and_then(Value::as_str)
-            .map(str::trim)
-            .filter(|runner| !runner.is_empty())
-            .map(ToString::to_string);
-        let local_fallback = request
-            .and_then(|value| value.get("local_fallback"))
-            .and_then(parse_local_fallback_policy);
-
-        AgentTaskLoopRunnerPolicy {
-            runner,
-            local_fallback,
-        }
-    }
-
     pub fn resolve_action_runner_policy<F>(
         &self,
         action: &AgentTaskLoopPolicyAction,
@@ -1071,7 +1023,7 @@ impl AgentTaskLoopControllerRecord {
     where
         F: FnMut(&str) -> AgentTaskLoopRunnerAvailability,
     {
-        let policy = self.runner_policy_for_action(action);
+        let policy = runner_policy_for_action(action);
         let fallback = policy.local_fallback.unwrap_or_else(|| {
             if policy.runner.is_some() {
                 AgentTaskLoopLocalFallbackPolicy::Denied
@@ -1802,60 +1754,6 @@ fn action_dedupe_key(action: &AgentTaskLoopPolicyAction) -> Option<String> {
             feedback_id.as_deref().unwrap_or("latest")
         )),
         _ => None,
-    }
-}
-
-fn action_runner_request(action: &AgentTaskLoopPolicyAction) -> Option<&Value> {
-    match action {
-        AgentTaskLoopPolicyAction::SpawnTask { request, .. } => Some(request),
-        AgentTaskLoopPolicyAction::FanOut {
-            request_template, ..
-        }
-        | AgentTaskLoopPolicyAction::RouteFinding {
-            request_template, ..
-        } => Some(request_template),
-        _ => None,
-    }
-}
-
-fn parse_local_fallback_policy(value: &Value) -> Option<AgentTaskLoopLocalFallbackPolicy> {
-    match value {
-        Value::Bool(true) => Some(AgentTaskLoopLocalFallbackPolicy::Allowed),
-        Value::Bool(false) => Some(AgentTaskLoopLocalFallbackPolicy::Denied),
-        Value::String(value) if value == "allowed" || value == "allow" || value == "true" => {
-            Some(AgentTaskLoopLocalFallbackPolicy::Allowed)
-        }
-        Value::String(value) if value == "denied" || value == "deny" || value == "false" => {
-            Some(AgentTaskLoopLocalFallbackPolicy::Denied)
-        }
-        _ => None,
-    }
-}
-
-fn blocked_runner_decision(
-    status: AgentTaskLoopActionStatus,
-    runner: Option<String>,
-    message: impl Into<String>,
-    details: Value,
-) -> AgentTaskLoopRunnerPolicyDecision {
-    AgentTaskLoopRunnerPolicyDecision {
-        target: None,
-        blocked_status: Some(status),
-        diagnostic: Some(AgentTaskLoopActionDiagnostic {
-            code: blocked_runner_status_code(status).to_string(),
-            message: message.into(),
-            runner,
-            details,
-        }),
-    }
-}
-
-fn blocked_runner_status_code(status: AgentTaskLoopActionStatus) -> &'static str {
-    match status {
-        AgentTaskLoopActionStatus::BlockedRunnerUnavailable => "blocked_runner_unavailable",
-        AgentTaskLoopActionStatus::BlockedRemoteMaterialization => "blocked_remote_materialization",
-        AgentTaskLoopActionStatus::BlockedLocalFallbackDenied => "blocked_local_fallback_denied",
-        _ => "runner_policy_not_blocked",
     }
 }
